gateway: Distribute hardware parent to other processes via UNIX sockets
Basically, any process that connects to /tmp/axum-gateway will
receive the hardware parent in string format, and the connection
will be closed immediately.
In the future, it might be a better idea to use a more universal
and flexible IPC protocol. (something similar to the protocol in
use by the new address server)