address: Start and end an SQL transaction when performing db locking
The lock was first meant to be used to prevent operations on the
SQL connection from multiple threads at the same time, internally.
But as the database is open for everyone to edit, this lock should
be extended to other applications accessing the DB as well. Luckily,
SQL has the concept of transactions, which handles this very nicely.
And luckily, we already had some sort of transactional interface
with the locking functions, so implementing this correctly is trivial.