engine: Added check for object count to prevent seg faults.
The engine determines the number of objects on the row count of the
template in the database. The practisch made clear the some objects
where not in the template list. Which gives a mismatch in object number
vs array number.  Now it is checked if we are larger then the row count number of objects.